+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## 0.22.4dev
 
+* [Fix] Fix error in `ScriptRunner` that didn't allow to import modules in script's directory (#1072)
+
 ## 0.22.3 (2023-04-13)
 
 * [Fix] Clearer error messages when generating pipeline plots
 * [Fix] Fix error when choosing which backend plot to use (when user didn't expicitly requested one)
-
 * [Fix] Fixed local execution error when using Scriptrunner
 
 ## 0.22.2 (2023-01-30)