+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## 0.22.4
+## 0.22.4dev
 
 * [Feature] More robust support for ploomber engine